move sbuf definitions to sbuf.h
